chore: show unsupported chain modal
